% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appigpt-package.R
\docType{package}
\name{appigpt-package}
\alias{appigpt}
\alias{appigpt-package}
\title{appigpt: gptstudio adds functionality based on OpenAI's GPT family of models directly into Rstudio}
\description{
OpenAI offers a number of large language models (LLMs) accessible via API. This package lowers the barrier to using this API inside of RStudio IDE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/appigpt}
  \item Report bugs at \url{https://github.com/PaulESantos/appigpt/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com}

}
\keyword{internal}
