% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_spg_impr.R
\name{spesgAddin}
\alias{spesgAddin}
\title{spelling and grammar Addin}
\usage{
spesgAddin()
}
\description{
Call this function as a Rstudio addin to ask GPT to improve spelling and
grammar of selected text.
}
