% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spesgAddin.R
\name{spesgAddin}
\alias{spesgAddin}
\title{Spanish spelling and grammar Addin}
\usage{
spesgAddin()
}
\description{
Call this function as a Rstudio addin to ask GPT to improve spanish spelling and grammar of selected text.
}
